Invert the priority order of the input and the rust-toolchain file
1. Not doing so would be a breaking change for users who have
a rust-toolchain file with a different version from what they
specify in the input.
2. Not doing so makes it overly complicated to install a specific
version.
Imagine you want to test your software against the version in the
rust-toolchain file and nightly as-well. Without this patch, users
will have the replace the content of the rust-toolchain file because
it is prioritized over the input.
It makes a lot more sense to prioritize the input over the file
because that one is easier changed.